+++ govee_api_laggat/__init__.py
@@ -1,3 +1,4 @@
+"""Govee API client package."""
 from .__version__ import VERSION
-from .govee_api_laggat import Govee, GoveeDevice
+from .govee_api_laggat import Govee, GoveeDevice, GoveeDeviceNotFound, GoveeError
 from .learning_storage import GoveeAbstractLearningStorage, GoveeLearnedInfo